refactor: invalid 한 class 및 method 이름 수정
